backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
pytest==7.4.3
pytest-cov==4.1.0
python-dotenv==1.0.1
psycopg[binary]==3.2.10
flask-socketio==5.3.6
pypdf2==3.0.1
openai==1.6.1
numpy>=1.24.0
setuptools>=65.0.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
